requirements.txt
fastapi
uvicorn
pillow
transformers
torch
python-multipart